+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 * Add department colours to components (PR #296)
+* Modify subscription links component (PR #294)
 
 # 7.1.0